code/lp_cal/requirements.txt
python-rtmidi==1.5.8
google-api-python-client
google-auth-httplib2
google-auth-oauthlib